requirements.txt

Flask==2.3.2
flask-cors==3.0.10
google-generativeai==0.4.1
